+++ lib/tools.php
